Tabs: change to AI, Create, Favs, Settings
