--- Sources/HumanRatedAI/ContentView.swift
@@ -1,74 +1,0 @@
-import SwiftUI
-
-public struct ContentView: View {
-    @AppStorage("tab") var tab = Tab.welcome
-    @AppStorage("name") var name = "Skipper"
-    @State var appearance = ""
-    @State var isBeating = false
-
-    public init() {
-    }
-
-    public var body: some View {
-        TabView(selection: $tab) {
-            VStack(spacing: 0) {
-                Text("Hello [\(name)](https://skip.tools)!")
-                    .padding()
-                Image(systemName: "heart.fill")
-                    .foregroundStyle(.red)
-                    .scaleEffect(isBeating ? 1.5 : 1.0)
-                    .animation(.easeInOut(duration: 1).repeatForever(), value: isBeating)
-                    .onAppear { isBeating = true }
-            }
-            .font(.largeTitle)
-            .tabItem { Label("Welcome", systemImage: "heart.fill") }
-            .tag(Tab.welcome)
-
-            NavigationStack {
-                List {
-                    ForEach(1..<1_000) { i in
-                        NavigationLink("Item \(i)", value: i)
-                    }
-                }
-                .navigationTitle("Home")
-                .navigationDestination(for: Int.self) { i in
-                    Text("Item \(i)")
-                        .font(.title)
-                        .navigationTitle("Screen \(i)")
-                }
-            }
-            .tabItem { Label("Home", systemImage: "house.fill") }
-            .tag(Tab.home)
-
-            NavigationStack {
-                Form {
-                    TextField("Name", text: $name)
-                    Picker("Appearance", selection: $appearance) {
-                        Text("System").tag("")
-                        Text("Light").tag("light")
-                        Text("Dark").tag("dark")
-                    }
-                    HStack {
-                        #if SKIP
-                        ComposeView { ctx in // Mix in Compose code!
-                            androidx.compose.material3.Text("💚", modifier: ctx.modifier)
-                        }
-                        #else
-                        Text(verbatim: "💙")
-                        #endif
-                        Text("Powered by \(androidSDK != nil ? "Jetpack Compose" : "SwiftUI")")
-                    }
-                    .foregroundStyle(.gray)
-                }
-                .navigationTitle("Settings")
-            }
-            .tabItem { Label("Settings", systemImage: "gearshape.fill") }
-            .tag(Tab.settings)
-        }
-        .preferredColorScheme(appearance == "dark" ? .dark : appearance == "light" ? .light : nil)
-    }
-}
-
-enum Tab : String, Hashable {
-    case welcome, home, settings
-}
